add some functions to include/psyq_3_0_missing/

+++ src/unknown/2539C.c
@@ -50,7 +50,7 @@ INCLUDE_ASM("asm/nonmatchings/unknown/2539C", Bresenham);
 /* score of ??? */
 /*INCLUDE_ASM("asm/nonmatchings/unknown/2539C", Bresenham);*/
 
-void Bresenham(void (*param_1)(s16, s16), s16 param_2, s16 param_3, s16 param_4, s16 param_5, s16 param_6, s16 param_7)
+void Bresenham(void (*param_1)(s32, s32), s16 param_2, s16 param_3, s16 param_4, s16 param_5, s16 param_6, s16 param_7)
 {
     s16 sp10;
     s16 sp12;
